Feat: 친구추가 기능 구현 완료 및 스케치북 전체 조회 #S10P31B103-118

+++ back/src/main/java/com/lemon/backend/domain/friend/service/impl/GroupsServiceImpl.java
@@ -1,6 +1,7 @@
 package com.lemon.backend.domain.friend.service.impl;
 
 import com.lemon.backend.domain.friend.dto.response.GroupResponseDto;
+import com.lemon.backend.domain.friend.entity.Friends;
 import com.lemon.backend.domain.friend.entity.Groups;
 import com.lemon.backend.domain.friend.repository.FriendsRepository;
 import com.lemon.backend.domain.friend.repository.GroupsRepository;
@@ -14,6 +15,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @Slf4j
@@ -45,9 +47,27 @@ public class GroupsServiceImpl implements GroupsService {
     }
 
     @Override
-    public void deleteGroup(Long groupId){
-        groupRepository.deleteById(groupId);
+    public void deleteGroup(Long groupId) {
+        Optional<Groups> groupOptional = groupRepository.findById(groupId);
+        if (groupOptional.isPresent()) {
+            Groups group = groupOptional.get();
+            List<Friends> friends = group.getFriendList(); // 혹시라도 속성 이름이 다르다면 적절히 변경해주세요.
+            // 기본 그룹 ID를 설정해주세요.
+            Long defaultGroupId = 1L;
+
+            Groups defaultGroup = groupRepository.findById(defaultGroupId).get();
+            for (Friends friend : friends) {
+                System.out.println(friend);
+                friend.setGroups(defaultGroup); // 속한 그룹을 기본 그룹으로 설정
+                friendsRepository.save(friend);
+            }
+            groupRepository.deleteById(groupId);
+        } else {
+            // 그룹이 존재하지 않는 경우 예외 처리
+            throw new CustomException(ErrorCode.INVALID_AUTH_CODE);
+        }
     }
+
 
     @Override
     public Long changeGroupName(Long groupId, String newName){
